minor clean ups

+++ src/main/java/goupb/webproject/Webproject/controller/ShoesController.java
@@ -1,8 +1,15 @@
 package goupb.webproject.Webproject.controller;
 import goupb.webproject.Webproject.dto.ShoesDTO;
+import goupb.webproject.Webproject.dto.UserDTO;
+import goupb.webproject.Webproject.entity.UserRole;
 import goupb.webproject.Webproject.service.ShoesService;
 
+import goupb.webproject.Webproject.service.UserService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -14,6 +21,9 @@ import java.util.Optional;
 public class ShoesController {
 
     private ShoesService shoesService;
+
+    @Autowired
+    private UserService userService;
 
     public ShoesController(ShoesService shoesService) {
         this.shoesService = shoesService ;
@@ -39,7 +49,11 @@ public class ShoesController {
 
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<Void> delete(@PathVariable String id){
-        shoesService.delete(id);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Optional<UserDTO> userDTO = userService.findByUsername(authentication.getName());
+        UserDTO user = userDTO.get();
+        if(user.getRole().equals(UserRole.ADMIN))shoesService.delete(id);
+        else throw new AccessDeniedException("Only admins can delete");
         return ResponseEntity.status(204).build();
     }
     @PutMapping("/update")
